Atualização da área restrita
