+++ src/controllers/authController.js
@@ -1,6 +1,8 @@
+const User = require('../models/User');
+
 const bcrypt = require('bcrypt');
 
-const User = require('../models/User');
+const saltRounds = 10;
 
 const userController = {
     //executa o login do usuário
@@ -27,12 +29,19 @@ const userController = {
         }
 
         // Se o email e a senha forem válidos, cria uma sessão para o usuário
-        // Salvando o email e o id do usuário na sessão
-        req.session.user = { email: usuario.email, id: usuario.id };
+        // Salva todas as informações do usuário na sessão
+        req.session.user = { 
+            email: usuario.email, 
+            id: usuario.id, 
+            nome: usuario.nome,
+            endereco: usuario.endereco,
+            cep: usuario.cep,
+            foto: usuario.foto
+        };
 
         // Redireciona para a página restrita
-        // Utiliza a rota userPrivate > /restrito
-        return res.redirect('restrito');
+        // Utiliza a rota userPrivate
+        return res.redirect('minha-conta');
     },
 
     executeUserLogout: (req, res) => {
@@ -43,23 +52,73 @@ const userController = {
         return res.redirect('/');
     },
 
-    renderLogin: (req, res) => {
+    renderUserFormLogin: (req, res) => {
         // Verifica se o usuário está logado
         // Ou seja, se existe uma sessão para o usuário
         if (req.session.user != undefined) {
           // Se estiver logado, redireciona para a página restrita
-          return res.redirect('restrito');
+          return res.redirect('minha-conta');
         }
     
-        // Renderiza a página de login
+        // se não estiver logado, redireciona para a página de login
         return res.render('pages/users/login',  { error: null });
     },
+
+    //renderiza a pagina de lista de usuarios
+    renderUserList: (req, res) => {
+        //busca todos os usuários no Db
+        const users = User.findAll();
+
+        //renderiza a página com lista de usuários
+        res.render('pages/users/list', { users });
+    },
     
+    //renderiza a página restrita (área logada do usuário)
     renderAreaRestrita: (req, res) => {
         // Busca o usuário na sessão
         const user = req.session.user;
+
         // Renderiza a página restrita passando os dados do usuário logado
-        return res.render('pages/users/areaRestrita',  { user });
+        return res.render('pages/users/areaUserLogado',  { user });
+    },
+
+    //renderiza página de perfil do usuário
+    renderUserPerfil: (req, res) => {
+        const user = req.session.user;
+        // const { id } = req.params;
+        // const user = User.findById(id);
+        res.render('pages/users/perfil', { user });
+    },
+
+    //renderiza dados do usuário na página de edição de cadastro
+    renderUserEditData: (req, res) => {
+        const { id } = req.params;
+        const user = User.findById(id);
+        res.render('pages/users/perfilEdit', { user });
+    },
+
+    //executa a atualização do cadastro do usuário
+    executeUserUpdate: (req, res) => {
+        const { id } = req.params;
+        const { nome, email, senha, cep, endereco, complemento } = req.body;
+        const foto = req.file.filename;
+        const hash = bcrypt.hashSync(senha, saltRounds);
+
+        User.removeFoto(id);
+        User.update(id, {nome, email, senha: hash, cep, endereco, complemento}, foto);
+        
+        //redireciona para home
+        res.redirect('/');
+    },
+
+    executeUserDelete: (req, res) => {
+        const { id } = req.params;
+
+        User.removeFoto(id);
+        User.delete(id);
+
+        //redireciona pra home
+        res.redirect('/');
     },
 }
