--- src/views/pages/users/areaRestrita.ejs
@@ -1,47 +1,0 @@
-<!-- PÁGINA TESTE PARA LOGIN E SESSION -->
-
-<html>
-  <!-- <head> -->
-    <%- include('../../partials/header') %>
-  <!-- </head> -->
-  
-  <body>
-    <style>
-      main{
-        padding: 30px;
-      }
-      
-      button{
-        margin-top: 10px;
-      }
-
-      h2{
-        font-size: 45px;
-        padding: 30px 0 30px 0;
-      }
-      
-    </style>
-
-    <main>
-    <h2> PÁGINA TEMPORÁRIA</h2>
-    <h1> Área Restrita </h1>
-    <p> Bem vindo a área restrita, <u>somente usuários logados podem ver essa página!</u> </p>
-     
-    <!-- Mostra os dados da sessão do usuário logado -->
-    <p> Nome do usuário: <%= user.nome %> </p>
-    <p> Email do Usuário: <%= user.email %> </p>
-    <p> Id do Usuário: <%= user.id %> </p>
-    
-    <a href="/users/lista">
-      <button>Lista de usuários</button>
-    </a>    
-
-    <!-- Formulário aponta para a rota de logout utilizando o método POST -->
-    <form action="/users/logout" method="POST">
-      <button>Fazer Logout</button>
-    </form>
-
-    </main>
-    
-  </body>
-</html>
